Function wasm_bindgen_test::__rt::__wasm_bindgen_generated___wbgtest_console_log

pub extern "C" fn __wasm_bindgen_generated___wbgtest_console_log(
    arg0: <Array as RefFromWasmAbi>::Abi
) -> <() as ReturnWasmAbi>::Abi

Handler for console.log invocations.

If a test is currently running it takes the args array and stringifies it and appends it to the current output of the test. Otherwise it passes the arguments to the original console.log function, psased as original.