Struct wasm_bindgen_test::__rt::Context

pub struct Context { /* private fields */ }

Runtime test harness support instantiated in JS.

The node.js entry script instantiates a Context here which is used to drive test execution.

Implementations

impl Context

pub fn new() -> Context

Creates a new context ready to run tests.

A Context is the main structure through which test execution is coordinated, and this will collect output and results for all executed tests.

pub fn args(&mut self, args: Vec<JsValue>)

Inform this context about runtime arguments passed to the test harness.

Eventually this will be used to support flags, but for now it’s just used to support test filters.

pub fn run(&self, tests: Vec<JsValue>) -> Promise

Executes a list of tests, returning a promise representing their eventual completion.

This is the main entry point for executing tests. All the tests passed in are the JS Function object that was plucked off the WebAssembly.Instance exports list.

The promise returned resolves to either true if all tests passed or false if at least one test failed.

impl Context

pub fn execute_sync<T: Termination>( &self, name: &str, f: impl 'static + FnOnce() -> T )

Entry point for a synchronous test in wasm. The #[wasm_bindgen_test] macro generates invocations of this method.

pub fn execute_async<F>(&self, name: &str, f: impl FnOnce() -> F + 'static)where F: Future + 'static, F::Output: Termination,

Entry point for an asynchronous in wasm. The #[wasm_bindgen_test(async)] macro generates invocations of this method.