Module wasm_bindgen_test::__rt

Internal-only runtime module used for the wasm_bindgen_test crate.

No API contained in this module will respect semver, these should all be considered private APIs.

Modules

browser
Support for printing status information of a test suite in a browser.

detect
Runtime detection of whether we’re in node.js or a browser.

node
Support for printing status information of a test suite in node.js

Structs

Context
Runtime test harness support instantiated in JS.

Functions

__wasm_bindgen_generated___wbgtest_console_debug
Handler for console.debug invocations. See above.

__wasm_bindgen_generated___wbgtest_console_error
Handler for console.error invocations. See above.

__wasm_bindgen_generated___wbgtest_console_info
Handler for console.info invocations. See above.

__wasm_bindgen_generated___wbgtest_console_log
Handler for console.log invocations.

__wasm_bindgen_generated___wbgtest_console_warn
Handler for console.warn invocations. See above.

__wbgtest_console_debug
Handler for console.debug invocations. See above.

__wbgtest_console_error
Handler for console.error invocations. See above.

__wbgtest_console_info
Handler for console.info invocations. See above.

__wbgtest_console_log
Handler for console.log invocations.

__wbgtest_console_warn
Handler for console.warn invocations. See above.

log
Internal implementation detail of the console_log! macro.